fix: oauth link 시, 기존 계정 없는 경우 예외 처리

+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/auth/usecase/OauthUseCase.java
@@ -63,6 +63,9 @@ public class OauthUseCase {
         phoneVerificationMapper.isValidCode(PhoneVerificationDto.VerifyCodeReq.from(request), PhoneVerificationType.getOauthSignUpTypeByProvider(provider));
         Pair<Boolean, String> isSignUpUser = checkSignUpUserNotOauthByProvider(provider, request.phone());
 
+        if (isSignUpUser.getLeft().equals(Boolean.FALSE) && request.username() == null)
+            throw new OauthException(OauthErrorCode.INVALID_OAUTH_SYNC_REQUEST);
+
         OidcDecodePayload payload = oauthOidcHelper.getPayload(provider, request.idToken());
         User user = userOauthSignMapper.saveUser(request, isSignUpUser, provider, payload.sub());
         phoneVerificationService.delete(request.phone(), PhoneVerificationType.getOauthSignUpTypeByProvider(provider));
